feat: add Perplexity API provider support (#760)
## Summary

Added Perplexity AI as a new provider in Bifrost, enabling access to Perplexity's Sonar models through the unified API.

## Changes

- Added Perplexity provider implementation with support for chat completions and responses
- Integrated Perplexity-specific features including search results, citations, and videos in response objects
- Added API key configuration in CI workflows for testing
- Updated documentation and UI to include Perplexity as a supported provider
- Updated README to reflect the increased provider count (now 15+)

## Type of change

- [x] Feature
- [ ] Bug fix
- [ ] Refactor
- [ ] Documentation
- [ ] Chore/CI

## Affected areas

- [x] Core (Go)
- [x] Transports (HTTP)
- [x] Providers/Integrations
- [ ] Plugins
- [x] UI (Next.js)
- [x] Docs

## How to test

1. Configure a Perplexity API key in your environment:

```sh
export PERPLEXITY_API_KEY=your_api_key_here
```

2. Run the core provider tests:

```sh
cd tests/core-providers
go test -v -run TestPerplexity
```

3. Test with the HTTP transport:

```sh
# Start Bifrost with the Perplexity provider enabled
./bifrost-http

# Make a request to the Perplexity provider
curl -X POST http://localhost:8000/v1/chat/completions \
  -H "Content-Type: application/json" \
  -d '{
    "model": "perplexity/sonar-pro",
    "messages": [{"role": "user", "content": "What is Bifrost?"}]
  }'
```

## Breaking changes

- [x] No
- [ ] Yes

## Related issues

N/A

## Security considerations

The implementation requires a Perplexity API key to be configured, which is handled securely through environment variables.

## Checklist

- [x] I added/updated tests where appropriate
- [x] I updated documentation where needed
- [x] I verified builds succeed (Go and UI)
- [x] I verified the CI pipeline passes locally if applicable

+++ core/providers/cohere/types.go
@@ -300,21 +300,21 @@ const (
 type CohereUsage struct {
 	BilledUnits  *CohereBilledUnits `json:"billed_units,omitempty"`  // Billed usage information
 	Tokens       *CohereTokenUsage  `json:"tokens,omitempty"`        // Token usage details
-	CachedTokens *float64           `json:"cached_tokens,omitempty"` // Cached tokens
+	CachedTokens *int               `json:"cached_tokens,omitempty"` // Cached tokens
 }
 
 // CohereBilledUnits represents billed usage information
 type CohereBilledUnits struct {
-	InputTokens     *float64 `json:"input_tokens,omitempty"`    // Number of billed input tokens
-	OutputTokens    *float64 `json:"output_tokens,omitempty"`   // Number of billed output tokens
-	SearchUnits     *float64 `json:"search_units,omitempty"`    // Number of billed search units
-	Classifications *float64 `json:"classifications,omitempty"` // Number of billed classification units
+	InputTokens     *int `json:"input_tokens,omitempty"`    // Number of billed input tokens
+	OutputTokens    *int `json:"output_tokens,omitempty"`   // Number of billed output tokens
+	SearchUnits     *int `json:"search_units,omitempty"`    // Number of billed search units
+	Classifications *int `json:"classifications,omitempty"` // Number of billed classification units
 }
 
 // CohereTokenUsage represents detailed token usage
 type CohereTokenUsage struct {
-	InputTokens  *float64 `json:"input_tokens"`  // Number of input tokens used
-	OutputTokens *float64 `json:"output_tokens"` // Number of output tokens produced
+	InputTokens  *int `json:"input_tokens"`  // Number of input tokens used
+	OutputTokens *int `json:"output_tokens"` // Number of output tokens produced
 }
 
 // CohereLogProb represents log probability information
